Encapsulates information about a conda environment

The CondaEnv class provides a basic structure to store information about a conda environment for a given method. It is designed to be used by the CondaEnvManager class.

Slots

method

1-length character. Indicating the method this CondaEnv refers to

needs.conda

logical. Whether the method needs conda (yes if python-based, no if R-based)

conda.bin

single string. path to conda binary. If empty (""), NULL or "conda", "auto" is passed to conda_binary to find path to conda binary in PATH

conda.env.name

1-length character. Name of the conda environment

conda.env.path

1-length character. Path to the conda environment

is.valid

logical. Whether the environment is valid. Should not be set by the user. Set by checkCondaEnv.

See also