Handy CondaEnvManger instance modifier

Wrapper to update the CondaEnvManager object stored in cache and used as configuration for conda environments for all known methods (typically stored in the package's cache). A single method can be updated at the time. In most cases, this is the best way to modify the package's conda environments registry.

Usage

UpdateEnvCache(
  method = known.methods,
  conda.bin = "auto",
  conda.env = NULL,
  conda.env.is.path = "auto",
  separate.scvi.envs = FALSE,
  overwrite.env = FALSE,
  dry.run = FALSE
)

Arguments

method

the name of the method to update. One of "combat", "harmony", "mnn", "bbknn", "scvi", "scanvi", "scanorama", "trvae"

conda.bin

the path to the conda binary. If empty (""), NULL or "conda", "auto" is passed to conda_binary() to find the path to the conda binary in PATH

conda.env

Either the name of a conda environment or the path to such an environment. Must be reachable by provided conda.bin. NULL enable to use the default environment names. (see Details section)

conda.env.is.path

Whether the conda.env is a path (TRUE), or a name (FALSE). The default ("auto") guesses. Be careful not to make a mistake if you switch to a non-default value

separate.scvi.envs

By default, SCVI and SCANVI share the same conda environment, since they rely on the same python package. If you wish to have a distinct environment for each, set it to TRUE. Ignored if method is not SCVI nor SCANVI.

overwrite.env

Turn it to TRUE to enable overwriting an existing environment (same name or same path). When the provided conda.env already exists, the default behaviour is to update the package's registry with the existing environment as it is

dry.run

When TRUE, the package's current cache is not updated. But the new conda environment (if any) will be created. FALSE by default.

Value

a CondaEnvManager object

Details

The conda environments default names are:

bbknn: SeuratIntegrate_scvi

SCVI: SeuratIntegrate_scvi-tools

SCANVI: SeuratIntegrate_scvi-tools

scanorama: SeuratIntegrate_scanorama

trVAE: SeuratIntegrate_trvae

See also